+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - [feat: add delay interface in tsf-example.](https://github.com/Tencent/spring-cloud-tencent/pull/1668)
 - [fix: fix lb configuration on bootstrap step.](https://github.com/Tencent/spring-cloud-tencent/issues/1673)
 - [feat:support 2024.0.2.](https://github.com/Tencent/spring-cloud-tencent/issues/1674)
+- [feat:support fault injection.](https://github.com/Tencent/spring-cloud-tencent/pull/1672)